+++ webviz_subsurface/_providers/__init__.py
@@ -1,2 +1,10 @@
+from .ensemble_summary_provider.ensemble_summary_provider import (
+    EnsembleSummaryProvider,
+    Frequency,
+    VectorMetadata,
+)
+from .ensemble_summary_provider.ensemble_summary_provider_factory import (
+    EnsembleSummaryProviderFactory,
+)
 from .ensemble_table_provider import EnsembleTableProvider, EnsembleTableProviderSet
 from .ensemble_table_provider_factory import EnsembleTableProviderFactory